Post component added.
Created the post component. Added an id for the parts that are going to experience a specific behavior.

Figured out that by default the <body> tag has a default margin of 8px, which is why my border wasn't reaching across the entire screen. Kinda annoying.

Once I'm done getting the page to scroll with fixed headers and footers, then I can look into building the Post page to finish the user flow. Exciting.